+++ bin/seed/seed-events.js
@@ -1,26 +1,9 @@
 'use strict';
 
 const mongoose = require('mongoose');
+const data = require('../../data/events.js');
 
 const Event = ('./models/events.js');
-
-const data = [{
-  name: 'Ironbeers',
-  type: 'Party',
-  date: '2018-09-14'
-}, {
-  name: 'BBQ',
-  type: 'Party',
-  date: '2018-08-31'
-}, {
-  name: 'Game Show',
-  type: 'Presentation',
-  date: '2018-09-07'
-}, {
-  name: 'UX-UI',
-  type: 'Talk',
-  date: '2018-09-12'
-}];
 
 mongoose.connect('mongodb://localhost/webbcn0818v2')
   .then(() => {
