+++ data/events.js
@@ -1,0 +1,17 @@
+const data = [{
+  name: 'Ironbeers',
+  type: 'Party',
+  date: '2018-09-14'
+}, {
+  name: 'BBQ',
+  type: 'Party',
+  date: '2018-08-31'
+}, {
+  name: 'Game Show',
+  type: 'Presentation',
+  date: '2018-09-07'
+}, {
+  name: 'UX-UI',
+  type: 'Talk',
+  date: '2018-09-12'
+}];
