+++ public/stylesheets/style.css
@@ -130,4 +130,3 @@ section.links {
   padding: 20px;
   border-radius: 5px;
 }
-
